+++ src/jsrepr.c
@@ -5,10 +5,10 @@
 #include "utf.h"
 #include "jscompile.h"
 
-extern void fmtnum(js_State *J, js_Buffer **sb, double n);
-static void reprvalue(js_State *J, js_Buffer **sb);
+extern void fmtnum(js_State *J, js_StringBuffer **sb, double n);
+static void reprvalue(js_State *J, js_StringBuffer **sb);
 
-static void reprnum(js_State *J, js_Buffer **sb, double n)
+static void reprnum(js_State *J, js_StringBuffer **sb, double n)
 {
 	char buf[40];
 	if (n == 0 && signbit(n))
@@ -17,7 +17,7 @@ static void reprnum(js_State *J, js_Buffer **sb, double n)
 		js_puts(J, sb, jsV_numbertostring(J, buf, n));
 }
 
-static void reprstr(js_State *J, js_Buffer **sb, const char *s)
+static void reprstr(js_State *J, js_StringBuffer **sb, const char *s)
 {
 	static const char *HEX = "0123456789ABCDEF";
 	Rune c;
@@ -54,7 +54,7 @@ static void reprstr(js_State *J, js_Buffer **sb, const char *s)
 #define isdigit(c) (c >= '0' && c <= '9')
 #endif
 
-static void reprident(js_State *J, js_Buffer **sb, const char *name)
+static void reprident(js_State *J, js_StringBuffer **sb, const char *name)
 {
 	const char *p = name;
 	if (isdigit(*p))
@@ -69,7 +69,7 @@ static void reprident(js_State *J, js_Buffer **sb, const char *name)
 		reprstr(J, sb, name);
 }
 
-static void reprobject(js_State *J, js_Buffer **sb)
+static void reprobject(js_State *J, js_StringBuffer **sb)
 {
 	const char *key;
 	int i, n;
@@ -100,7 +100,7 @@ static void reprobject(js_State *J, js_Buffer **sb)
 	js_putc(J, sb, '}');
 }
 
-static void reprarray(js_State *J, js_Buffer **sb)
+static void reprarray(js_State *J, js_StringBuffer **sb)
 {
 	int n, i;
 
@@ -127,7 +127,7 @@ static void reprarray(js_State *J, js_Buffer **sb)
 	js_putc(J, sb, ']');
 }
 
-static void reprfun(js_State *J, js_Buffer **sb, js_Function *fun)
+static void reprfun(js_State *J, js_StringBuffer **sb, js_Function *fun)
 {
 	int i;
 	js_puts(J, sb, "function ");
@@ -141,7 +141,7 @@ static void reprfun(js_State *J, js_Buffer **sb, js_Function *fun)
 	js_puts(J, sb, ") { [byte code] }");
 }
 
-static void reprvalue(js_State *J, js_Buffer **sb)
+static void reprvalue(js_State *J, js_StringBuffer **sb)
 {
 	if (js_isundefined(J, -1))
 		js_puts(J, sb, "undefined");
@@ -230,7 +230,7 @@ static void reprvalue(js_State *J, js_Buffer **sb)
 
 void js_repr(js_State *J, int idx)
 {
-	js_Buffer *sb = NULL;
+	js_StringBuffer *sb = NULL;
 	int savebot;
 
 	if (js_try(J)) {